# api/documentation/swagger_resources/shops-api-doc.json
{
   "swagger":"2.0",
   "host":"api.kortingscode.nl",
   "basePath":"",
   "tags":[
      {
         "name":"shop",
          "description":"Shop Operations"
      }
   ],
   "schemes":[
      "http"
   ],
   "paths":{
      "/shops":{
         "post":{
            "tags":[
               "shop"
            ],
            "summary":"Add a new shop",
            "description":"",
            "operationId":"addshop",
            "consumes":[
               "application/json"
            ],
            "produces":[
               "application/json"
            ],
            "parameters":[
               {
                  "in":"body",
                  "name":"body",
                  "description":"Shop that needs to be insert",
                  "required":true,
                  "schema":{
                     "$ref":"#/definitions/shop"
                  }
               }
            ],
            "responses":{
               "200":{
                  "description":"Request was successful"
               },
               "405":{
                  "description":"Invalid input"
               }
            }
         }
      },

      "/shops/{shopId}":{
         "put":{
            "tags":[
               "shop"
            ],
            "summary":"Update an existing shop",
            "description":"",
            "operationId":"updateshop",
            "consumes":[
               "application/json"
            ],
            "produces":[
               "application/json"
            ],
            "parameters":[
               {
                     "in":"path",
                     "name":"shopId",
                     "description":"Shop Id that needs to be update",
                     "required":true,
                     "type":"integer",
                     "format":"int64"
               },
               {
                  "in":"body",
                  "name":"body",
                  "description":"Shop that needs to be update",
                  "required":true,
                  "schema":{
                     "$ref":"#/definitions/shop"
                  }
               }
            ],
            "responses":{
               "200":{
                  "description":"Request was successful"
               },
               "404":{
                  "description":"Shop not found"
               },
               "405":{
                  "description":"Invalid input"
               }
            }
         },
         "get":{
            "tags":[
               "shop"
            ],
            "summary":"Find shop by ID",
            "description":"Returns a single shop",
            "operationId":"getshopById",
            "produces":[
               "application/json"
            ],
            "parameters":[
                 {
                     "in":"path",
                     "name":"shopId",
                     "description":"Shop that needs to be fetched",
                     "required":true,
                     "type":"integer",
                     "format":"int64"
                 }
             ],
            "responses":{
               "200":{
                  "description":"Request was successful",
                  "schema":{
                     "$ref":"#/definitions/shop"
                  }
               },
               "404":{
                  "description":"Shop not found"
               },
               "400":{
                  "description":"Invalid shop Id"
               }
            }
         },
         "delete":{
            "tags":[
               "shop"
            ],
            "summary":"Deletes a shop",
            "description":"",
            "operationId":"deleteshop",
            "produces":[
               "application/json"
            ],
            "parameters":[
                 {
                     "in":"path",
                     "name":"shopId",
                     "description":"Shop that needs to be deleted",
                     "required":true,
                     "type":"integer",
                     "format":"int64"
                 }
             ],
             "responses":{
                 "200":{
                     "description":"Request was successful"
                 },
                 "404":{
                     "description":"Shop not found"
                 },
                 "400":{
                     "description":"Invalid shop Id"
                 }
             }
         }
      }
     
   },
   "definitions":{
      "shop":{
         "type":"object",
         "required":[
            "name",
            "navigationUrl"
         ],
         "properties":{
            "name":{
               "type":"string",
               "example":"MyShop"
            },
            "permaLink":{
               "type":"string",
               "example":"http://flipit.com/in/ebay"
            },
            "overriteTitle":{
               "type":"string",
               "example":""
            },
            "metaDescritption":{
               "type":"string",
               "example":""
            },
            "usergenratedcontent":{
               "type":"int",
               "description":"Allow user generated content",
               "enum":[1,0]
            },
            "discussions":{
               "type":"int",
               "description":"Allow discussions",
               "enum":[1,0]
            },
            "title":{
               "type":"string",
               "example":""
            },
            "subTitle":{
               "type":"string",
               "example":""
            },
            "notes":{
               "type":"string",
               "example":""
            },
            "accountManagerName":{
               "type":"string",
               "description":"Exact name 'first name'",
               "example":""
            },
            "affliateNetwork":{
               "type":"string",
               "description":"",
               "example":""
            },
            "deepLinkStatus":{
               "type":"int",
               "description":"Deeplink code (Yes/No)",
               "enum":[1,0]
            },
            "refUrl":{
               "type":"string",
               "description":"Ref Url(Page slug)"
            },
            "actualUrl":{
               "type":"string",
               "description":"Actual Url",
               "example":"http://www.test.com"
            },
            "shoptext":{
               "type":"string"
            },
            "classification":{
                "type":"string",
                "enum": [
                    "C",
                    "B",
                    "A",
                    "A+",
                    "AA",
                    "AA+",
                    "AAA"
                ]
            }
         }
      },
      "ApiResponse":{
         "type":"object",
         "properties":{
            "code":{
               "type":"integer",
               "format":"int32"
            },
            "type":{
               "type":"string"
            },
            "message":{
               "type":"string"
            }
         }
      }
   }
}